+++ editor_patch/main.cpp
@@ -505,6 +505,7 @@ void ApplyTriggerPatches();
 void ApplyLevelPatches();
 void ApplyEventsPatches();
 void ApplyTexturesPatches();
+void ApplyLightmapPatches();
 
 void LoadAlpineEditorPackfile()
 {
@@ -729,9 +730,6 @@ extern "C" DWORD AF_DLL_EXPORT Init([[maybe_unused]] void* unused)
     InitLogging();
     InitCrashHandler();
 
-    // stops pink lightmaps, but also crashes with too many lights todo: can this be fixed?
-    //AsmWriter{0x004AC60B}.jmp(0x004AC611);
-
     // Apply AF-specific changes only if legacy mode isn't active
     if (get_is_saving_af_version()) {
         apply_af_level_editor_changes();
@@ -785,6 +783,7 @@ extern "C" DWORD AF_DLL_EXPORT Init([[maybe_unused]] void* unused)
     ApplyLevelPatches();
     ApplyEventsPatches();
     ApplyTexturesPatches();
+    ApplyLightmapPatches();
 
     // Browse for .v3m files instead of .v3d
     static char mesh_ext_filter[] = "Mesh (*.v3m)|*.v3m|All Files (*.*)|*.*||";